+++ src/components/createForm/createEntityForm/EntityForm.css
@@ -1,0 +1,23 @@
+/* .scroller-row {
+    display: flex;
+    flex-grow: 1;
+    flex-wrap: wrap;
+    max-width: 100%;
+    overflow: hidden;
+} */
+
+.selected-users-container {
+    max-height: 120px;
+    /* Adjust the height as needed */
+    overflow-y: auto;
+}
+
+.user-list {
+    max-height: 150px;
+    /* Adjust the height as needed */
+    overflow-y: auto;
+    list-style: none;
+    padding: 0;
+    margin: 0;
+ 
+}
