+++ src/components/pages/settings/SettingsComponents/Communication/Email.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Email = () => {
+  return (
+    <div>Email</div>
+  )
+}
+
+export default Email
