+++ src/components/pages/settings/SettingsComponents/Communication/Whatsapp.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const Whatsapp = () => {
+  return (
+    <div>Whatsapp</div>
+  )
+}
+
+export default Whatsapp
