+++ src/components/pages/settings/SettingsComponents/Roles/Roles.jsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import { Link } from 'react-router-dom';
+
+const Roles = () => {
+    return (
+        <div className='container p-4 bg-[#f8fafc]'>
+            <h1 className='mx-3 font-semibold text-lg grid1-item'>Roles</h1>
+
+        </div>
+    )
+}
+
+export default Roles
